Working, but very hacky progressUpdater
